Moving initialize scripts
Moving all scripts to do with initialize that aren't used directly in the pipeline to their own folder
