+++ dataactcore/scripts/initialize.py
@@ -13,21 +13,21 @@ from dataactcore.broker_logging import configure_logging
 from dataactcore.models.userModel import User
 from dataactcore.models.jobModels import FileGeneration
 from dataactcore.models.validationModels import RuleSetting
-from dataactcore.scripts.load_agencies import load_agency_data
-from dataactcore.scripts.load_cfda_data import load_cfda_program
-from dataactcore.scripts.load_country_codes import load_country_codes
-from dataactcore.scripts.load_defc import load_defc
-from dataactcore.scripts.load_funding_opportunity_number import load_funding_opportunity_number_data
-from dataactcore.scripts.load_location_data import load_location_data
-from dataactcore.scripts.load_object_class import load_object_class
-from dataactcore.scripts.setup_all_db import setup_all_db
-from dataactcore.scripts.setup_emails import setup_emails
-from dataactcore.scripts.load_submission_window_schedule import load_submission_window_schedule
-from dataactcore.scripts.load_tas import load_tas
+from dataactcore.scripts.initialize.load_agencies import load_agency_data
+from dataactcore.scripts.initialize.load_cfda_data import load_cfda_program
+from dataactcore.scripts.initialize.load_country_codes import load_country_codes
+from dataactcore.scripts.initialize.load_defc import load_defc
+from dataactcore.scripts.initialize.load_funding_opportunity_number import load_funding_opportunity_number_data
+from dataactcore.scripts.initialize.load_location_data import load_location_data
+from dataactcore.scripts.initialize.load_object_class import load_object_class
+from dataactcore.scripts.initialize.setup_all_db import setup_all_db
+from dataactcore.scripts.initialize.setup_emails import setup_emails
+from dataactcore.scripts.initialize.load_submission_window_schedule import load_submission_window_schedule
+from dataactcore.scripts.initialize.load_tas import load_tas
+from dataactcore.scripts.initialize.read_zips import read_zips
 from dataactcore.scripts.pipeline.load_program_activity import load_program_activity_data
 from dataactcore.scripts.pipeline.load_sf133 import load_all_sf133
 from dataactcore.scripts.pipeline.load_tas_failing_edits import load_all_tas_failing_edits
-from dataactcore.scripts.read_zips import read_zips
 
 from dataactvalidator.health_check import create_app
 from dataactvalidator.filestreaming.labelLoader import LabelLoader
