+++ dataactcore/scripts/initialize/load_location_data.py
