+++ dataactcore/scripts/initialize/setup_static_data.py
