+++ dataactcore/scripts/initialize/setup_submission_type_db.py
