+++ dataactcore/scripts/initialize/setup_user_db.py
